naoTrainer/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
deepface==0.0.93
face-alignment==1.4.1
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.1.3
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
pyrealsense2==2.56.4.9191
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
retina-face==0.0.17
rich==14.0.0
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.0
shiboken6==6.9.1
six==1.17.0
sounddevice==0.5.2
soupsieve==2.7
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
tifffile==2025.5.10
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
